chore: refactor pr
